+++ mesh/src/main/java/ca/mcmaster/cas/se2aa4/a2/mesh/adt/services/Thickenable.java
@@ -7,11 +7,11 @@ public interface Thickenable {
      *
      * @param x The new {@link Color} to set to the object
      */
-    void setThickness(double x);
+    void setThickness(float x);
 
     /**
      *
      * @return The {@link Color} assigned to the object
      */
-    double getThickness();
+    float getThickness();
 }
